benchmarks/VideoMarkBench/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
av==14.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
clip==0.2.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
eagerpy==0.30.0
easydict==1.13
einops==0.8.0
ffmpeg-python==0.2.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.55.3
foolbox==3.3.4
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
invisible-watermark==0.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
llvmlite==0.43.0
lpips==0.1.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
metrics==0.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.9
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
open_clip_torch==2.29.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.1
pycocotools==2.0.8
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-msssim==1.0.0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
requests==2.32.3
rich==14.0.0
robustml==0.0.3
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.9.20
timm==1.0.12
tokenizers==0.21.0
torch==2.6.0
torch-dct==0.1.6
torchmetrics==1.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.47.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.28.post3
yarl==1.18.3
